requirements.txt
Flask==3.0.3
azure-cosmos==4.7.0
python-dotenv==1.0.1
